ID: #9 #3 Edit: Add navigation function for boards in BoardService.js class
